libls/evloop_lfuncs.h: replace spinlock with mutex
Using spinlocks in userspace is a bad idea, outside of a couple of
very specialized workloads (which have nothing to do with what we do
here). The reason is, they are subject to the so-called "spinning
while preempted" problem; see
https://www.realworldtech.com/forum/?threadid=189711&curpostid=189723
for more info. The solution is to use mutexes (mutices?), which do not
hide the "dependency graph" from the kernel.

It is actually quite absurd that we have to use mutexes to perform
what can be done atomically on virtually any 64-bit machine; but such
is the price of being standard-compliant and portable, I guess.

(Interestingly, we would not even be able to use C11 atomics here: C11
just completely ignores floating-point stuff when describing _Atomic.)