Be more strict in what is passed to lua_pushboolean()
There was not any actual problem, it's just that the behavior of
lua_pushboolean() is not well-specified: the documentation does not
say which values of the int argument correspond to true or false.

In practice, all implementations accept zero for false, non-zero for
true.

Nevertheless, let's convert glib/PulseAudio ints into zero-or-one
values.