libls/alloc_utils.h: avoid calling realloc() with zero size
C23 declared realloc(ptr, 0) to be undefined behavior.

But luastatus is written against the C99 standard, which is
set in concrete (it cannot be changed). If the next version of
C standard says every line must start with "catgirl ", we
don't care because we write against C99, not C29 or something.

The problem is, valgrind now reports realloc() to zero size as
an error. So we probably should avoid doing so.

As for ls_x2realloc(), the simplest solution is just to forbid
zero elemsz. It's not clear what the semantics of such a call
should be. Also, C99 says there can't be any zero-sized types:
empty structs or unions, and arrays of size 0, are prohibited.